SF: Limit frame rate overrides to 60 Hz if touch boost is disabled
In Android U "RefreshRateSelector::getFrameRateOverrides()" added
support for non-native refresh rate which enabled frame rate overrides
as low as 30 Hz. This works nice as long as touch boost is also
enabled. On systems without touch boost this causes noticable
non-smooth scrolling. The particular usecase where this could
be seen was in scrolling YouTube search results when 30-FPS thumbnail
video was playing. The solution to the problem is to keep 60 Hz as
lowest limit on systems without touch boost.

Bug: 323048643
Test: atest libsurfaceflinger_unittest
Depends-On: I5ca238f832ec630e0b86de9ef6e61fb3a5b86f5b
Change-Id: Idc3684cdd77485f0d806805d7b019b8c873b1e00